Added `#' @export` just before start of function definition in 'project_z_by_z.R'.
This is needed to make sure the function is noted in the file 'NAMESPACE' as exported. When the package is built only exported functions are made available to the user after `library(ccmppWPP)`. If the function is not exported the user would have to call the function using `ccmppWPP:::project_z_by_z(...)`.