Tidy up test files. (started... wip)
Started to move creation of objects from 'helper_' to the actual test files. This is done because the creation of the objects triggers quite a lot of other function calls that are then tested again later.. not really getting rid of the duplicate testing (?) but slighlty less confusing if the objects are created only in the test files.