Validation of `ccmpp_input_list`s no longer (re)validates all elements.
Added argument `.validate_elements` to the validation method and set to `FALSE` when used in side the creator function `ccmpp_input_list`. All elements are validated earlier in the function so it is unnecessary to validate their contents again. Checks are still done on their attributes etc.

Time saved is about 50%. Over 5 replications,

BEFORE

 user.self   sys.self    elapsed user.child  sys.child
     2.642      0.040      2.688        NaN        NaN

AFTER

 user.self   sys.self    elapsed user.child  sys.child
     1.366      0.008      1.368        NaN        NaN