(failing tests); `aggregate()` method replaced with `collapse_demog_dimension()`; `abridge()` needs working on
Following the principle applied to other existing generics such as `subset()`, the `aggregate()` method was removed and replaced with a new custom function `collapse_demog_dimension()`. This is easier to understand and more consistent with the existing approach. `collapse_demog_dimension()` is basically the same as the `aggregate()` method was.

The `abridge()` function needs to be tested after these modifications.