Archive the C++ version of CCMPP. See 'inst/archive/ccmpp_cpp'.
Why? Including C++ code requires a lot of extra installation steps and these introduce risks of errors or incompatibility. The C++ version is not used anywhere in the workflow so this is not really worth the risk at this stage. The motivation for the C++ version was to increase speed mainly for use with probabilistic reconstruction, but currently that uses an implementation of CCMPP written in STAN. If, at a later date, we figure out how to link C++ functions to STAN models, then the C++ version here could be used as a basis for that.