Develop/mark (#30)
* Very minor change.

Change `T` to `TRUE`

* Tiny addition to documentation

* Add argument `n_max` to all calls to `readxl::read_xlsx()` in 'ccmppWPP_input_file.R'.

Eagle input files contain Excel Tables and it _appears_ that `readxl::read_xlsx()` uses these to determine the cell range to import. This caused problems because the Tables are currently not reliably extended when the input files are modified via the Eagle interface. Specifying `n_max` should hopefully force the all rows to be read. `readxl::read_xlsx()` supposedly then discards empty rows (without regard to whether or not they are in a Table).

* Add warning message about Albanian census years

* Note use of `sort_demog_change_component_df()` in `popReconstruct2` package.

* Tick mark breaks on x-axes of plots set to '10'.

* Improve x-axis breaks selection for plots

* Update README.md

* Documentation update

* Update docs and metadata

* Update README

* Update README

* Update README.md

* Update README.md

* Updates to package infra after merging branch 'develop/sara'

* Fix bug

* Delete messages about setting `value_type` and `value_scale` if either are supplied as `NULL`

These messages are mostly just a nuisance. These arguments are almost always `NULL` and the selection of default values is reasonably responsive to different inputs so these are not really needed. The selection of defaults has been detailed in the documentation to `demog_change_component_df()`.

* Hide some warnings in test script and make mig rate thresholds sensitive to value scale

* `collapse_demog_dimension()` only works on `demog_change_component_df`s

* Improve warning messages from `validate_ccmppWPP_object.mig_net_rate_age_sex()`

* Add `importFrom` roxygen tag to make sure all the functions needed by `ccmppWPP_aggregate()` from tidyverse are imported.

* - Add a check to `DDextract_ccmppWPPinputs_tier1()` to make sure data source years are valid.
- Fix error when there are no missing SRB values.
- In tests, ipdate arguments to calls to `DDextract_ccmppWPPinputs_tier1()`. New data source years.

* Commit new NAMESPACE

* Fix tidyverse "deprecated" message.

`ccmppWPP_compute_WPP_outputs()` used `summarise()` with an output of more than one row. Replaced `summarise` with `reframe`

* Update some tests to prevent messages and warnings popping up.

* Fix doc errors (mostly unbalanced braces).

* Add a tolerance to checking if age-cohort and age-period deaths are the same.

---------

Co-authored-by: Mark Wheldon <wheldon@un.org>
Co-authored-by: Mark C Wheldon <markalava@users.noreply.github.com>